+++ app/pages/TrendsRegion.jsx
@@ -77,7 +77,7 @@ class TrendsRegion extends Component {
                                 </List>
                             </Cell>
                             <Cell col={10}>
-                                <div className={styles.rightmap}>
+                                <div className={styles.rightMap}>
                                     <Map
                                         display_draw='False'
                                         trends_settings={trendsPageSettings}