+++ app/utils/getConfig.js
@@ -1,6 +1,7 @@
 /*
  * @flow
  */
+
 import type { PropertiesType } from './flowtype';
 
 export function getSourceName(source:PropertiesType):string {
@@ -141,3 +142,24 @@ export function getErrorText() {
     return window.configruntime.gd3.error_text;
 }
 
+export function getLayersDetails() {
+    let layersDetails = [];
+    if (window.configruntime.gd3.exploreLayers){
+        layersDetails = window.configruntime.gd3.exploreLayers;
+    }
+    return layersDetails;
+}
+
+export function getAvailableLayers() {
+    let available_layers = [];
+
+    getLayersDetails().map(p =>
+        available_layers.push({
+            'title': p.properties.title,
+            'opacity': p.properties.opacity,
+            'visibility': p.properties.visibility,
+        })
+    );
+
+    return available_layers;
+}